refactor: Restructure query code
